Unwind circular dependency
