+++ Cargo.toml
@@ -41,6 +41,7 @@ members = [
     "programs/storage_program",
     "programs/token_api",
     "programs/token_program",
+    "programs/utilities",
     "programs/vote_api",
     "programs/vote_program",
     "replicator",
