+++ genesis/src/main.rs
@@ -2,7 +2,7 @@
 
 use clap::{crate_description, crate_name, crate_version, value_t_or_exit, App, Arg};
 use solana::blocktree::create_new_ledger;
-use solana_runtime::genesis_utils::get_default_native_instruction_processors;
+use solana_program_utilities::get_default_native_instruction_processors;
 use solana_sdk::account::Account;
 use solana_sdk::fee_calculator::FeeCalculator;
 use solana_sdk::genesis_block::Builder;
