+++ programs/utilities/Cargo.toml
@@ -1,0 +1,55 @@
+[package]
+name = "solana-program-utilities"
+version = "0.18.0-pre0"
+description = "Solana program utilities"
+authors = ["Solana Maintainers <maintainers@solana.com>"]
+repository = "https://github.com/solana-labs/solana"
+license = "Apache-2.0"
+homepage = "https://solana.com/"
+edition = "2018"
+
+[dependencies]
+# bincode = "1.1.4"
+# bv = { version = "0.11.0", features = ["serde"] }
+# byteorder = "1.3.2"
+# fnv = "1.0.6"
+hashbrown = "0.2.0"
+# lazy_static = "1.3.0"
+# libc = "0.2.58"
+# libloading = "0.5.2"
+# log = "0.4.7"
+# maplit = "1.0.1"
+# memmap = "0.6.2"
+# rand = "0.6.5"
+# rayon = "1.1.0"
+# serde = "1.0.97"
+# serde_derive = "1.0.97"
+# serde_json = "1.0.40"
+# solana-logger = { path = "../../logger", version = "0.18.0-pre0" }
+# solana-measure = { path = "../../measure", version = "0.18.0-pre0" }
+# solana-metrics = { path = "../../metrics", version = "0.18.0-pre0" }
+solana-bpf-loader-api = { path = "../bpf_loader_api", version = "0.18.0-pre0" }
+solana-bpf-loader-program = { path = "../bpf_loader_program", version = "0.18.0-pre0" }
+solana-budget-api= { path = "../budget_api", version = "0.18.0-pre0" }
+solana-budget-program = { path = "../budget_program", version = "0.18.0-pre0" }
+solana-config-api = { path = "../config_api", version = "0.18.0-pre0" }
+solana-config-program = { path = "../config_program", version = "0.18.0-pre0" }
+solana-exchange-api = { path = "../exchange_api", version = "0.18.0-pre0" }
+solana-exchange-program = { path = "../exchange_program", version = "0.18.0-pre0" }
+solana-move-loader-program = { path = "../move_loader_program", version = "0.18.0-pre0" }
+solana-move-loader-api = { path = "../move_loader_api", version = "0.18.0-pre0" }
+solana-sdk = { path = "../../sdk", version = "0.18.0-pre0" }
+solana-stake-api = { path = "../stake_api", version = "0.18.0-pre0" }
+solana-stake-program = { path = "../stake_program", version = "0.18.0-pre0" }
+solana-storage-api = { path = "../storage_api", version = "0.18.0-pre0" }
+solana-storage-program = { path = "../storage_program", version = "0.18.0-pre0" }
+solana-token-api = { path = "../token_api", version = "0.18.0-pre0" }
+solana-token-program = { path = "../token_program", version = "0.18.0-pre0" }
+solana-vote-api = { path = "../vote_api", version = "0.18.0-pre0" }
+solana-vote-program = { path = "../vote_program", version = "0.18.0-pre0" }
+
+# sys-info = "0.5.7"
+
+[lib]
+crate-type = ["lib"]
+name = "solana_program_utilities"
